chore: Decrease verbosity of logging for unknown webhook kinds
The current verbosity overwhelms the logs thanks to issue and status
hooks, which we don't currently handle. Since we just use the exact
same hook configuration on GitHub as Prow would, we're getting those
whether we want them or not, and with #545 eventually we'll want
them. But for now, let's only vomit out all that noise on debug.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>