fix: Set a pending commit status on contexts as we launch them
This is to avoid the annoying delay between the metapipeline being
created and the `PipelineActivity` existing and being reported by the
build controller.

More importantly, it'll also allow us to block merges from happening
while there's a required presubmit pending but not yet "building",
which I've seen. This is because we check the contexts on the PR to
see if they're passing before merging. Previously, when rerunning a PR
because we've just approved it and master had changed since it passed
(so a rebuild is neccessary), there'd be a delay between Tide kicking
off the rebuild and when the GitHub commit status would actually say
the new rebuild was running. If Tide synced again during that delay,
it would see that the statuses were fine and would merge. Which is
wrong. Now, it'll see pending contexts and not merge.

All told, this is yet another reason reporting needs to be separated
from the build controller and we need an intermediary layer like `ProwJob`.

fixes #562
fixes #563

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>