fix: Add support for GitHub App authentication to webhook controller
This takes how Tide was already doing this and makes it reusable, with
some tweaks for the webhook controller needing to already have an
instantiated scm client for webhook parsing, before it knows the
owner, and makes sure we're following the same code path for scm
provider client creation and authorization everywhere.

Also, this reworks the charts a bit to include the relevant secret and
configuration in the webhook controller deployment as well.

fixes #586

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>