feat: Preliminary support for providers without PR labels via comments
This change specifically adds support for this in `scmprovider.Client`
via magic, but doesn't do anything with Keeper yet, since that works
under the assumption that the labels field will already be populated
in PR query results. That'll have to be reworked to update the PRs
individually when appropriate, but I want to get this in first.

part of #727

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>