fix: Support requiring for contexts with conditional execution
Using the `runIfChanged` JX scheduler field for pre- and post-submits
(`run_if_changed` in the Prow/Lighthouse configuration for presubmits
that we actually generate from the jx scheduler configuration) and the
`alwaysRun` jx scheduler field, you can specify a regex in
`runIfChanged` and `false` in `alwaysRun`, in which case the context
will only run when the list of files changed in the PR matches the regex.

That's been in place for a while, but until this change, the jx
scheduler->config generation would get things confused and make any
context specified in `policy.requiredStatusContexts` be treated as
required, whether it's automatically run/conditionally run/etc. This
is largely because we kinda did things a bit dumb in how we generate
the branch protection required status checks. Prow has a cronjob that
will update GitHub branch protection rules based on the "branch
protection" configuration in Prow, but we've never used that with jx,
and we don't support it in Lighthouse, at least not at this point, but
since I don't want to rule out the possibility in the future, I'd like
to leave this janky setup as is.

So now we will remove anything from the "required" list if it's also
in the "required if present" list, just to be safe.

fixes #435

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>