chore: First attempt at making jx "optional"
I don't care for this, but I think it should work for the moment. Once
I actually write a Tekton implementation, it should be more clear. I
opted to not split off into a separate controller for the moment and
instead just have the existing ones be aware of whether to do JX stuff
or not, with this change just determining how we start the controllers up.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>