wip: Rewrite a bunch of logic in ci.sh
Should fix GC of clusters and nginx validation webhook error, and now
will error out if any of the required env vars is missing. We also
switch to using gomplate for templating the values file rather than a
bunch of sed junk, and generally clean the bash up a bunch.

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>