feat: Add ability to autoincrement build numbers
Currently available only for Tekton. It'll start at 1 for each
combination of `(org)-(repo)-(branch or PR)-(context)`. The "history"
is persisted in a configmap - there may be a better option/approach.

This is heavily derived from Prow's `tot` service.

Still needs to be worked into the e2e tests.

fixes #950

Signed-off-by: Andrew Bayer <andrew.bayer@gmail.com>