fix: bigger default memory limit for webhooks
the `webhooks` component needs to run the `git clone` command once for each repository. If you have big repositories, this could fail if you don't give enough memory to the container: the OOMKiller will kill the git process.
Let's used a "better" default value, and add a comment for people having the issue.

fixes:#1085

Signed-off-by: Vincent Behar <v.behar@free.fr>